QWEN2.5_42_7b_main/requirements.txt
fire
lz4
numpy
matplotlib
pandas
pytz
seaborn
scipy
transformers
torch
huggingface_hub
tokenizers
safetensors
regex
tqdm
requests
urllib3
certifi
charset-normalizer
idna
filelock
fsspec
jinja2
markupsafe
pyyaml
accelerate
psutil
pillow
cycler
kiwisolver
pyparsing
python-dateutil
fonttools
contourpy
packaging
six
fire